Move _staging out of model
It needs to access stuff the model module does need to touch

JIRA: RAIL-3383
