+++ libs/sdk-ui-dashboard/.dependency-cruiser.js
@@ -5,19 +5,20 @@ options = {
         ...depCruiser.DefaultRules,
         ...depCruiser.DefaultSdkRules,
         depCruiser.moduleWithDependencies("drill", "src/drill", [
+            "src/_staging/*",
             "src/constants",
             "src/dashboardAux",
             "src/localization",
             "src/model",
-            "src/model/_staging/*",
             "src/types.ts",
         ]),
         depCruiser.moduleWithDependencies("filterBar", "src/filterBar", [
+            "src/_staging/*",
             "src/localization",
             "src/model",
-            "src/model/_staging/*",
         ]),
         depCruiser.moduleWithDependencies("insight", "src/insight", [
+            "src/_staging/*",
             "src/converters",
             "src/dashboardAux",
             "src/drill",
@@ -25,23 +26,22 @@ options = {
             "src/localization",
             "src/logUserInteraction",
             "src/model",
-            "src/model/_staging/*",
             "src/types.ts",
         ]),
         depCruiser.moduleWithDependencies("kpi", "src/kpi", [
+            "src/_staging/*",
             "src/converters",
             "src/dashboardAux",
             "src/dashboardItems",
             "src/drill",
             "src/localization",
             "src/model",
-            "src/model/_staging/*",
             "src/types.ts",
         ]),
         depCruiser.moduleWithDependencies("model", "src/model", [
+            "src/_staging/*",
             "src/converters",
             "src/drill/types.ts",
-            "src/layout/DefaultDashboardLayoutRenderer/index.ts", // TODO RAIL-3383 remove this after _staging moves outside of model
             "src/localization",
             "src/types.ts",
         ]),
