+++ libs/sdk-ui-dashboard/src/_staging/dateFilterConfig/dateFilterOptionMapping.ts
