+++ libs/sdk-ui-dashboard/src/_staging/dateFilterConfig/defaultConfig.ts
