+++ libs/sdk-ui-dashboard/src/_staging/dateFilterConfig/sanitization.ts
