+++ libs/sdk-ui-dashboard/src/_staging/dateFilterConfig/tests/fixtures.ts
