+++ libs/sdk-ui-dashboard/src/_staging/dateFilterConfig/tests/sanitization.test.ts
