+++ libs/sdk-ui-dashboard/src/_staging/dateFilterConfig/tests/validation.test.ts
