+++ libs/sdk-ui-dashboard/src/_staging/insight/insightBuilder.ts
