+++ libs/sdk-ui-dashboard/src/_staging/note.md
