+++ libs/sdk-ui-dashboard/src/model/commandHandlers/dashboard/loadDashboardHandler/index.ts
@@ -26,8 +26,8 @@ import { internalErrorOccurred } from "../../../events/general";
 import { loadUser } from "./loadUser";
 import { userActions } from "../../../state/user";
 import { metaActions } from "../../../state/meta";
-import { dashboardFilterContextDefinition } from "../../../_staging/dashboard/dashboardFilterContext";
-import { dashboardLayoutSanitize } from "../../../_staging/dashboard/dashboardLayout";
+import { dashboardFilterContextDefinition } from "../../../../_staging/dashboard/dashboardFilterContext";
+import { dashboardLayoutSanitize } from "../../../../_staging/dashboard/dashboardLayout";
 import { loadDashboardList } from "./loadDashboardList";
 import { listedDashboardsActions } from "../../../state/listedDashboards";
 import { backendCapabilitiesActions } from "../../../state/backendCapabilities";
