+++ libs/sdk-ui-dashboard/src/model/commandHandlers/dashboard/loadDashboardHandler/resolveDashboardConfig.ts
@@ -1,27 +1,28 @@
 // (C) 2021 GoodData Corporation
-import {
-    DashboardConfig,
-    DashboardContext,
-    isResolvedConfig,
-    ResolvedDashboardConfig,
-} from "../../../types/commonTypes";
+import { SagaIterator } from "redux-saga";
+import { all, call } from "redux-saga/effects";
 import {
     IDateFilterConfigsQueryResult,
     ISeparators,
     ISettings,
     IUserWorkspaceSettings,
 } from "@gooddata/sdk-backend-spi";
-import { LoadDashboard } from "../../../commands/dashboard";
-import { SagaIterator } from "redux-saga";
-import { all, call } from "redux-saga/effects";
-import { dateFilterValidationFailed } from "../../../events/dashboard";
-import { defaultDateFilterConfig } from "../../../_staging/dateFilterConfig/defaultConfig";
-import { dispatchDashboardEvent } from "../../../eventEmitter/eventDispatcher";
-import { IColorPalette } from "@gooddata/sdk-model";
-import { PromiseFnReturnType } from "../../../types/sagas";
 import { ILocale } from "@gooddata/sdk-ui";
-import { getValidDateFilterConfig } from "../../../_staging/dateFilterConfig/validation";
-import { stripUserAndWorkspaceProps } from "../../../_staging/settings/conversion";
+import { IColorPalette } from "@gooddata/sdk-model";
+
+import { defaultDateFilterConfig } from "../../../../_staging/dateFilterConfig/defaultConfig";
+import { getValidDateFilterConfig } from "../../../../_staging/dateFilterConfig/validation";
+import { stripUserAndWorkspaceProps } from "../../../../_staging/settings/conversion";
+import { LoadDashboard } from "../../../commands/dashboard";
+import { dateFilterValidationFailed } from "../../../events/dashboard";
+import { dispatchDashboardEvent } from "../../../eventEmitter/eventDispatcher";
+import {
+    DashboardConfig,
+    DashboardContext,
+    isResolvedConfig,
+    ResolvedDashboardConfig,
+} from "../../../types/commonTypes";
+import { PromiseFnReturnType } from "../../../types/sagas";
 
 function loadDateFilterConfig(ctx: DashboardContext): Promise<IDateFilterConfigsQueryResult | undefined> {
     const { backend, workspace } = ctx;
